requirements.txt
# FastAPI 및 ASGI 서버
fastapi
gunicorn

# Google Cloud SDK
google-cloud-storage
google-cloud-spanner
google-auth
google-api-python-client
google-auth-oauthlib

# 인증 및 HTTP 요청
requests
oauthlib
requests-oauthlib
aiohttp

# 유틸리티
markdown
python-dotenv
